requirements.txt
Flask>=2.3.2
opencv-python>=4.8.0
pyserial>=3.5
pytest>=7.4.0
ultralytics>=8.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pyyaml>=6.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
psutil>=5.9.0